### 2019-03-31 #1357 @danielskatz
👋 @azazel7 - there's a problem with your paper, as seen from the whedon comment above.  Please make changes, following the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography), and then enter ```@whedon generate pdf``` as a new comment here to recompile the paper, and repeat until you are happy with it.
