### 2019-03-31 #1357 @danielskatz
In this case, the affiliations in the author block need to be simple integers that refer to the affiliations block, not complete affiliations in the author block.  Again, see the example.
