### 2019-03-31 d992d4e
Fix issue in paper.md.
