### 2019-04-14 #1357 @labarba
@azazel7 — We're missing in your paper a brief description of how the software is used or can be used in research settings. Could you expand that? Please also comment on any extant software that targets the same (or similar) application or functionality.
