### 2019-04-23 6b3b931
Add references to MOA and WEKA, two frameworks similar to OrpailleCC.
