+++ paper/paper.bib
@@ -135,3 +135,22 @@
   journal={White paper},
   year={2016}
 }
+@article{moa,
+  title={Moa: Massive online analysis},
+  author={Bifet, Albert and Holmes, Geoff and Kirkby, Richard and Pfahringer, Bernhard},
+  journal={Journal of Machine Learning Research},
+  volume={11},
+  number={May},
+  pages={1601--1604},
+  year={2010}
+}
+@article{weka,
+  title={The WEKA data mining software: an update},
+  author={Hall, Mark and Frank, Eibe and Holmes, Geoffrey and Pfahringer, Bernhard and Reutemann, Peter and Witten, Ian H},
+  journal={ACM SIGKDD explorations newsletter},
+  volume={11},
+  number={1},
+  pages={10--18},
+  year={2009},
+  publisher={ACM}
+}
