### 2019-04-24 #1357 @azazel7
A paragraph has been refactored to add the required information. The goal of OrpailleCC is to provide a consistent environment to study data stream algorithms for embedded systems. I will use this library for my own research as well.
