### 2019-04-26 #1357 @danielskatz
@azazel7 - At some point, you will need to add DOIs to references that have them.  Once you have done so, you can regenerate the PDF by adding a new comment here that is ```@whedon generate pdf``` (not as part of another comment, but as a new comment)
