### 2019-04-26 #1357 @danielskatz
see the above for suggestions, but be aware, it is not always completely correct, so take it with a grain of salt...
