### 2019-04-26 aa98f5b
Add missing DOIs. Remove duplicate references. Adjust paper.tex to compile properly.
