+++ paper/paper.tex
@@ -4,6 +4,7 @@
 \usepackage{graphicx}
 \usepackage{float}
 \usepackage{caption}
+\usepackage{textcomp}
 \usepackage{xcolor}
 \usepackage[lofdepth,lotdepth]{subfig}
 
@@ -21,11 +22,17 @@ applications could leverage the computing power of connected objects
 for improved scalability.
 
 OrpailleCC provides a consistent collection of data stream algorithms 
-developed to be deployed on embedded devices. Its main objective
-is to enable programmers of embedded systems to use 
-out-of-the-box algorithms with an efficient implementation. It also 
-intends to facilitate the comparison and benchmarking of algorithms in a 
-consistent framework.
+developed to be deployed on embedded devices.  Its main objective is to
+support research on data stream mining for connected objects,
+by facilitating the comparison and benchmarking of algorithms in a 
+consistent framework. It also enables programmers of embedded systems to use 
+out-of-the-box algorithms with an efficient implementation.
+To the best of our knowledge, existing libraries of stream mining
+algorithms cannot be used on connected objects due to their resource consumption or
+assumptions about the target system (e.g., existence of a `malloc` function).
+Nevertheless, for more powerful devices such as desktop computers, Java
+frameworks such as Massive Online Analysis~\cite{moa} and WEKA~\cite{weka} achieve
+similar goals as OrpailleCC.
 
 OrpailleCC targets the classes of problems discussed in~\cite{kejariwal2015},
  in particular Sampling and 
@@ -48,7 +55,7 @@ multi-dimensional variant~\cite{ltcd} are two methods to compress data
 streams.
 
 All implementations rely as little as possible on functions provided by the 
-operating system, for instance \texttt{malloc}, since such functions are typically
+operating system, for instance `malloc`, since such functions are typically
 not available on embedded systems. When algorithms cannot be
 implemented without such functions, the library uses template parameters to 
 request the required functions from the user.  All algorithms are 
@@ -59,7 +66,8 @@ was chosen for its performance as well as its popularity in the
 field. All methods are tested and tests are run through Travis-CI.
 
 In the future, we plan to extend the library with other reliable 
-algorithms to widely cover as many common problems as possible.
+algorithms to widely cover as many common problems as possible. We also plan to 
+use it as a basis to design new stream classification methods.
 External contributions are, of course, most welcome.
 
 
