### 2019-04-26 #1357 @azazel7
@danielskatz I found most of the DOIs of the references in the paper, except for two of them.
