### 2019-04-27 #1357 @danielskatz
Great - for the couple of the conference proceedings where the cases are incorrect, you might add another {} around the text where you want to preserve the case in the .bib file.
