### 2019-04-29 bb1f24e
Fix case in bib file.
