+++ paper/paper.bib
@@ -9,7 +9,7 @@
 	pages = {2040--2041},
 	author = {Arun Kejariwal and Sanjeev Kulkarni and Karthik Ramasamy},
 	title = {Real time analytics},
-	journal = {Proceedings of the {VLDB} Endowment}
+	journal = {Proceedings of the {VLDB} {E}ndowment}
 }
 @inproceedings{ltc,
 	doi = {10.1109/lcn.2004.72},
@@ -17,7 +17,7 @@
 	publisher = {{IEEE} (Comput. Soc.)},
 	author = {T. Schoellhammer and B. Greenstein and E. Osterweil and M. Wimbrow and D. Estrin},
 	title = {Lightweight temporal compression of microclimate datasets [wireless sensor networks]},
-	booktitle = {29th Annual {IEEE} International Conference on Local Computer Networks}
+	booktitle = {29th Annual {IEEE} {I}nternational {C}onference on {L}ocal {C}omputer {N}etworks}
 }
 @inproceedings{ltcd,
 	doi = {10.1109/bigdata.2018.8621946},
@@ -27,15 +27,15 @@
 	publisher = {{IEEE}},
 	author = {Bo Li and Omid Sarbishei and Hosein Nourani and Tristan Glatard},
 	title = {A multi-dimensional extension of the Lightweight Temporal Compression method},
-	booktitle = {2018 {IEEE} International Conference on Big Data (Big Data)}
+	booktitle = {2018 {IEEE} {I}nternational {C}onference on {B}ig {D}ata (Big Data)}
 }
 @inproceedings{chained_reservoir_sampling,
   title={Sampling from a moving window over streaming data},
   author={Babcock, Brian and Datar, Mayur and Motwani, Rajeev},
-  booktitle={Proceedings of the thirteenth annual Association for Computing Machinery-SIAM symposium on Discrete algorithms},
+  booktitle={Proceedings of the thirteenth annual {A}ssociation for {C}omputing {M}achinery-{SIAM} {D}ymposium on {D}iscrete algorithms},
   pages={633--634},
   year={2002},
-  organization={Society for Industrial and Applied Mathematics}
+  organization={{S}ociety for {I}ndustrial and {A}pplied {M}athematics}
 }
 @inproceedings{cuckoo_filter,
 	doi = {10.1145/2674005.2674994},
@@ -44,7 +44,7 @@
 	publisher = {{ACM} Press},
 	author = {Bin Fan and Dave G. Andersen and Michael Kaminsky and Michael D. Mitzenmacher},
 	title = {Cuckoo Filter},
-	booktitle = {Proceedings of the 10th {ACM} International on Conference on emerging Networking Experiments and Technologies - {CoNEXT} {\textquotesingle}14}
+	booktitle = {Proceedings of the 10th {ACM} {I}nternational on {C}onference on emerging {N}etworking {E}xperiments and {T}echnologies - {CoNEXT} {\textquotesingle}14}
 }
 @article{mc-nn,
 	doi = {10.1016/j.future.2017.03.026},
@@ -56,27 +56,27 @@
 	pages = {187--199},
 	author = {Mark Tennant and Frederic Stahl and Omer Rana and Jo{\~{a}}o B{\'{a}}rtolo Gomes},
 	title = {Scalable real-time classification of data streams with concept drift},
-	journal = {Future Generation Computer Systems}
+	journal = {Future {G}eneration {C}omputer {S}ystems}
 }
 @article{reservoir_sampling,
 	doi = {10.1145/3147.3165},
 	url = {https://doi.org/10.1145%2F3147.3165},
 	year = 1985,
 	month = {mar},
[…]
