+++ paper/paper.bib
@@ -16,7 +16,7 @@
 	url = {https://doi.org/10.1109%2Flcn.2004.72},
 	publisher = {{IEEE} (Comput. Soc.)},
 	author = {T. Schoellhammer and B. Greenstein and E. Osterweil and M. Wimbrow and D. Estrin},
-	title = {Lightweight temporal compression of microclimate datasets [wireless sensor networks]},
+	title = {Lightweight {T}emporal {C}ompression of {M}icroclimate {D}atasets [wireless sensor networks]},
 	booktitle = {29th Annual {IEEE} {I}nternational {C}onference on {L}ocal {C}omputer {N}etworks}
 }
 @inproceedings{ltcd,
@@ -26,13 +26,13 @@
 	month = {dec},
 	publisher = {{IEEE}},
 	author = {Bo Li and Omid Sarbishei and Hosein Nourani and Tristan Glatard},
-	title = {A multi-dimensional extension of the Lightweight Temporal Compression method},
+	title = {A multi-dimensional extension of the {L}ightweight {T}emporal {C}ompression method},
 	booktitle = {2018 {IEEE} {I}nternational {C}onference on {B}ig {D}ata (Big Data)}
 }
 @inproceedings{chained_reservoir_sampling,
   title={Sampling from a moving window over streaming data},
   author={Babcock, Brian and Datar, Mayur and Motwani, Rajeev},
-  booktitle={Proceedings of the thirteenth annual {A}ssociation for {C}omputing {M}achinery-{SIAM} {D}ymposium on {D}iscrete algorithms},
+  booktitle={Proceedings of the thirteenth annual {A}ssociation for {C}omputing {M}achinery-{SIAM} {S}ymposium on {D}iscrete algorithms},
   pages={633--634},
   year={2002},
   organization={{S}ociety for {I}ndustrial and {A}pplied {M}athematics}
@@ -161,7 +161,7 @@
   year={2016}
 }
 @article{moa,
-  title={Moa: Massive online analysis},
+  title={{MOA}: {M}assive {O}nline {A}nalysis},
   author={Bifet, Albert and Holmes, Geoff and Kirkby, Richard and Pfahringer, Bernhard},
   journal={Journal of {M}achine {L}earning {R}esearch},
   volume={11},
