### 2019-05-02 #1357 @azazel7
@gkthiruvathukal : I had a look at the list and I would suggest the following names:
- federeghe
- sepandhaghighi
- aaronyd
- gcdeshpande
- conradsnicta
- adavidzh
- lisitsyn

I made sure that all have a background in C++ as well as knowledge on embedded systems and machine learning. None of them are known by me or any author of OrpailleCC.
