### 2019-05-19 #1357 @gkthiruvathukal
@federeghe Would you be willing to contribute a review for this JOSS submission?
