### 2019-05-21 6a89a6b
Fix issue with error_threshold. There is still some randomness in the split test.
