### 2019-05-22 dcbf004
Remove randomness from the test split of mc-nn.
The randomness could lead to a state of learning where the split was not
necessary.
