### 2019-05-22 3591d10
Modify find_nearest_clusters functions in mc-nn.
Now, when nearest and nearest_with_class are different cluster but have
the same distance, then nearest is replaced with nearest_with_class.
Because in such a case, the order of the clusters influence on which
cluster will be found first.
