### 2019-06-03 #1485 @whedon
[REVIEW]: OrpailleCC: a Library for Data Stream Analysis on Embedded Systems
**Submitting author:** @Azazel7 (<a href="http://orcid.org/0000-0003-2129-5517">Martin Khannouz</a>)
**Repository:** <a href="https://github.com/big-data-lab-team/OrpailleCC" target ="_blank">https://github.com/big-data-lab-team/OrpailleCC</a>
**Version:** 1.0

- [x] **Version:** 1.0
