### 2019-06-04 big-data-lab-team/OrpailleCC#5 @sepandhaghighi
JOSS Review : Installation
Please add more information about installation procedure and usage to `README.md`, something like minimum system requirements, dependencies, OS and etc.
