### 2019-06-04 big-data-lab-team/OrpailleCC#6 @sepandhaghighi
JOSS Review : Community guidelines
Add these items to your repo :

- `CHANGELOG.md` ([Example](https://keepachangelog.com/en/1.0.0/))
- `AUTHORS.md` ([Example](https://github.com/sepandhaghighi/pycm/blob/master/AUTHORS.md))
- `.github` folder ([Example](https://github.com/sepandhaghighi/pycm/tree/master/.github))
	- `CODE_OF_CONDUCT.md`
	- `CONTRIBUTING.md` (detailed)
	- `ISSUE_TEMPLATE.md`
	- `PULL_REQUEST_TEMPLATE.md`
