### 2019-06-04 big-data-lab-team/OrpailleCC#7 @sepandhaghighi
JOSS Review : Docstring
Please add docstring to **all** of your functions and methods.
