### 2019-06-04 big-data-lab-team/OrpailleCC#8 @sepandhaghighi
JOSS Review : Automated tests
Your project automated test is fine. But it would be better if you measure and report coverage of your tests.
