### 2019-06-04 big-data-lab-team/OrpailleCC#9 @sepandhaghighi
JOSS Review : A statement of need
I suggest you to update first section of `README.md` and talk a little more about your project target audience.
I also suggest you to add a logo.
