### 2019-06-04 big-data-lab-team/OrpailleCC#10 @sepandhaghighi
JOSS Review : Version
And finally after previous minor issues, please add this project to [Zenodo](https://zenodo.org/) and release the first version (V1.0).
