### 2019-06-04 #1485 @sepandhaghighi
review in progress, would say "accept with minor revisions".
added 6 issues, after those are closed, this should be ready to accept
big-data-lab-team/OrpailleCC/issues/5
big-data-lab-team/OrpailleCC/issues/6
big-data-lab-team/OrpailleCC/issues/7
big-data-lab-team/OrpailleCC/issues/8
big-data-lab-team/OrpailleCC/issues/9
big-data-lab-team/OrpailleCC/issues/10
