### 2019-06-07 big-data-lab-team/OrpailleCC#11 @sepandhaghighi
JOSS Review : Code analysis tools
Analyze your code by some proper tools, these tools helps to identify **vulnerability** and **bugs** in source code.

Recommended analyzers :

1. [Codacy](https://www.codacy.com/)
2. [Codefactor](https://www.codefactor.io/)
3. [LGTM](https://lgtm.com/)

also take a look at [Github Marketplace](https://github.com/marketplace?category=code-quality&type=apps).
