### 2019-06-12 d0f12ee
Add install instructions.
