### 2019-06-12 1d1c5ac
Add docstring to all functions.
