### 2019-06-12 5455315
Add coverage commands in the Makefile.
