### 2019-06-12 524d0b8
Add few line in the readme.
Add community files.
