+++ CHANGELOG.md
@@ -1,0 +1,14 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+### Added
+- Reservoir Sampling
+- Chained Reservoir Sampling
+- Micro-Cluster Nearest Neighbour (MC-NN)
+- Lightweight Temporal Compression (LTC)
+- Bloom Filter
+- Cuckoo Filter
