### 2019-06-12 638a60f
Add more docstring in mc_nn.hpp.
