### 2019-06-13 b7b5863
Change the scope of a variable in CuckooFilter.
