### 2019-06-13 a01b580
Fix a bug in MCNN::cluster::performance.
