### 2019-06-13 6c63468
Rename the directory for the coverage website.
