### 2019-06-13 9e10cca
Add a global header in the readme. Update the references.
