+++ README.md
@@ -1,3 +1,7 @@
+<div align="center">
+<hr/>
+<h1>OrpailleCC</h1>
+</div>	
 [![Codacy Badge](https://api.codacy.com/project/badge/Grade/0d8b194e7b1148f7a6bb3c0ba19c9309)](https://app.codacy.com/app/azazel7/OrpailleCC?utm_source=github.com&utm_medium=referral&utm_content=azazel7/OrpailleCC&utm_campaign=Badge_Grade_Dashboard)
 [![Build Status](https://travis-ci.org/azazel7/OrpailleCC.svg?branch=master)](https://travis-ci.org/azazel7/OrpailleCC)
 
@@ -9,9 +13,9 @@ framework. It also enables programmers of embedded systems to use
 out-of-the-box algorithms with an efficient implementation.
 Algorithms from OrpailleCC are based on C++ templates and does not use the STL library.
 
-# Get started
-## Hello World
-Let us run a basic example with a reservoir sampling [4] of size 3.
+## Get started
+### Hello World
+Let us run a basic example with a reservoir sampling \[4] of size 3.
 Save the following code in *testy.cpp*.
 ```cpp
 #include <iostream> //Included for cout
@@ -40,8 +44,8 @@ $ g++ -I./src -std=c++11 testy.cpp -o testy
 $ ./testy
 Hll
 ```
-## Install
-### Requirement
+### Install
+#### Requirement
 As the collection is designed to run on embedded system without operating
 systems, OrpailleCC has very little dependencies and requirement.
 
@@ -50,7 +54,7 @@ systems, OrpailleCC has very little dependencies and requirement.
 - googletest: to run unit tests.
 - Linux Operating System: because all instructions are given for Linux systems. However, OrpailleCC should compile properly on a Windows system as long as a C++ compiler is available.
 
-### Installation
+#### Installation
 To install OrpailleCC, first clone the repository. 
 ```bash
 git clone https://github.com/big-data-lab-team/OrpailleCC.git
@@ -92,18 +96,18 @@ cd ~/hello
 g++ -std=c++11 -I$ORPAILLECC_DIR hello.c
 ```
 
-## Test
-### Unit Test
+### Test
+#### Unit Test
 The unit tests require the `googletest` library ([Here](https://github.com/google/googletest)).
 To run the unit tests, run the command `make run_test`.
 
-### Performance
+#### Performance
 To run a performance test on your device, compile the performance tests with
 `make perf` then run `./main-perf`.
 
 ![Alt](/figures/performance.png "An example of the performance output")
 
-### Coverage
+#### Coverage
 To observe the coverage of test function, run the following commands:
 ```bash
 make clean
@@ -112,10 +116,10 @@ make config=debug coverage
 These commands will clean previous object files to rebuild them with the debug options, then run the test and gather the data for the coverage.
 To visualize the test coverage, simply open *coverage/index.html* into your favorite browser.
 
-# Examples
+## Examples
 This section provides the list of all algorithms implemented in OrpailleCC with a brief example.
-## Lightweight Temporal Compression (LTC)
-LTC [0] is a compression algorithm that approximates a series of values with a linear
+### Lightweight Temporal Compression (LTC)
+LTC \[0] is a compression algorithm that approximates a series of values with a linear
 function. The epsilon parameter controls the amount of compression. If the
 linear approximation isn't accurate enough, then a new point is
 issued.
@@ -139,8 +143,8 @@ int main(){
 	}
 }
 ```
-## Micro-Cluster Nearest Neighbour (MC-NN)
-MC-NN [3] is a classifier based on k-nearest neighbours. It aggregates the data
+### Micro-Cluster Nearest Neighbour (MC-NN)
[…]
