### 2019-06-13 b6f4b6b
Fix links to Travis and codacy.
