### 2019-06-18 5426203
Add coverage run in Travis.
