### 2019-06-20 3fd0c2d
Updated badges
