### 2019-06-20 big-data-lab-team/OrpailleCC#11 @azazel7
Fixed using Codacy.
