### 2019-06-20 big-data-lab-team/OrpailleCC#8 @azazel7
Fixed with gcov and lcov.
A section in the readme describes how to obtain the test coverage.
