### 2019-06-20 big-data-lab-team/OrpailleCC#5 @azazel7
A new section has been added in the readme to properly describe how to install and use OrpailleCC.
