### 2019-06-20 big-data-lab-team/OrpailleCC#6 @azazel7
The missing files have been added.
