### 2019-06-20 big-data-lab-team/OrpailleCC#9 @azazel7
The first section of the readme has been improved to better describe the project.
