### 2019-06-20 big-data-lab-team/OrpailleCC#10 @azazel7
Fixed with Zenodo. The badge has been added to the readme.
