### 2019-06-20 #1485 @azazel7
All issues should be fixed now.
