### 2019-06-20 big-data-lab-team/OrpailleCC#6 @sepandhaghighi
@azazel7 Thanks for your efforts!
Please consider the following :
1. Update your email address in `CODE_OF_CONDUCT.md`
2. Update `CONTRIBUTING.md`, test section!!
3. Update `AUTHORS.md` (other authors!!)
4. Add version(1.0) header in `CHANGELOG.md` ([Example](https://github.com/sepandhaghighi/pycm/blob/master/CHANGELOG.md))
