### 2019-06-20 #1485 @sepandhaghighi
> All issues should be fixed now.

Nice work, but some minor issues remained!

1. Please take another look at [this issue](https://github.com/big-data-lab-team/OrpailleCC/issues/6)
2. Missing DOIs
3. Solve paper compilation problem
