### 2019-06-21 7fe1267
Update email in CODE_OF_CONDUCT.
