+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+
+## [1.0] - 2019-06-20
+### Added
 - Reservoir Sampling
 - Chained Reservoir Sampling
 - Micro-Cluster Nearest Neighbour (MC-NN)
