### 2019-06-21 d926f58
Add lcov package in Travis for test coverage.
