### 2019-06-21 big-data-lab-team/OrpailleCC#6 @azazel7
The changes have been made.
