### 2019-06-21 53b762b
Try to fix header format in paper.md.
