### 2019-06-21 #1485 @azazel7
If I understand correctly, the file `paper.yml` is missing. However, there is no mention of such a file in the author guideline nor in the documentation at joss.readthedocs.io. I have checked the repositories of already published papers, but I couldn't find any example of a paper.yml.
