### 2019-06-22 #1485 @danielskatz
paper.yml is internally created - you do not have to create this. Perhaps this was a whedon hiccup, which we can check by trying again, or perhaps there is something wrong in your md or bib file that is leading to a problem with an error message that is not helpful
