### 2019-06-26 dbd8590
Revert changes to fix paper.md.
