+++ paper/paper.bib
@@ -160,25 +160,3 @@
   journal={White paper},
   year={2016}
 }
-@article{moa,
-  title={{MOA}: {M}assive {O}nline {A}nalysis},
-  author={Bifet, Albert and Holmes, Geoff and Kirkby, Richard and Pfahringer, Bernhard},
-  journal={Journal of {M}achine {L}earning {R}esearch},
-  volume={11},
-  number={May},
-  pages={1601--1604},
-  year={2010}
-}
-@article{weka,
-	doi = {10.1145/1656274.1656278},
-	url = {https://doi.org/10.1145%2F1656274.1656278},
-	year = 2009,
-	month = {nov},
-	publisher = {Association for {C}omputing {M}achinery ({ACM})},
-	volume = {11},
-	number = {1},
-	pages = {10},
-	author = {Mark Hall and Eibe Frank and Geoffrey Holmes and Bernhard Pfahringer and Peter Reutemann and Ian H. Witten},
-	title = {The {WEKA} data mining software},
-	journal = {{ACM} {SIGKDD} {E}xplorations {N}ewsletter}
-}
