### 2019-06-26 e6b2112
Rollback to fix pdf generation.
