### 2019-06-26 20111da
Remove all references (debug session).
