### 2019-06-26 22d0291
Make paper.md no whedon friendly (debug session). See if it can produce other errors.
