+++ paper/paper.md
@@ -15,8 +15,6 @@ authors:
    orcid: 0000-0003-2620-5883
    affiliation: 1
 affiliations:
- - name: Department of Computer Science and Software Engineering, Concordia University, Montreal, Canada
-   index: 1
 date: 29 March 2019
 bibliography: paper.bib
 ---
