### 2019-06-26 7d3c795
Fix affiliation count and indexes (debug version).
