### 2019-06-26 c09b3a7
Remove authors. (debug version)
