### 2019-06-26 edb8aee
Remove affiliation. (debug version)
