+++ paper/paper.md
@@ -8,10 +8,6 @@ authors:
  - name: Martin Khannouz
    orcid: 0000-0003-2129-5517
    affiliation: 1
-affiliations:
-- index: 1
-  name: Concordia University
-date: 29 March 2019
 bibliography: paper.bib
 ---
 
