### 2019-06-26 8657482
Add affiliation in paper.yml. (debug version)
Error message has changed this time.
