+++ paper/paper.yml
@@ -8,6 +8,10 @@ authors:
  - name: Martin Khannouz
    orcid: 0000-0003-2129-5517
    affiliation: 1
+affiliations:
+ - name: Department of Computer Science and Software Engineering, Concordia University, Montreal, Canada
+   index: 1
+date: 29 March 2019
 bibliography: paper.bib
 ---
 
