### 2019-06-26 5b18a73
Add keywords section in paper.yml. (debug version)
