+++ paper/paper.yml
@@ -4,6 +4,10 @@ tags:
  - C++
  - data stream
  - embedded systems
+keywords:
+ - C++
+ - data stream
+ - embedded systems
 authors:
  - name: Martin Khannouz
    orcid: 0000-0003-2129-5517
