### 2019-06-26 6e7ca07
Duplicate paper.md with its equivalent paper.yml. (debug version)
