### 2019-06-26 #1485 @sepandhaghighi
@azazel7
I suggest you, remove `paper.yml`, `paper.tex` and `Makefile` and work in wizard mode (only `paper.md` and `paper.bib`)
