+++ paper/Makefile
@@ -1,8 +1,0 @@
-all:
-	pdflatex paper.tex
-	bibtex paper
-	pdflatex paper.tex
-	pdflatex paper.tex
-
-clean:
-	rm -f paper.aux paper.bbl paper.blg paper.log paper.pdf
