### 2019-06-27 #1485 @sepandhaghighi
@azazel7 Now add your references to `paper.md`
