### 2019-06-27 c792be3
Reset paper.md and paper.bid to their previous content. (debug version)
