+++ paper/paper.bib
@@ -1,86 +1,114 @@
 @article{kejariwal2015,
-	author	= "Arun Kejariwal, Sanjeev Kulkarni, Karthik Ramasamy",
-	title	= "Real Time Analytics: Algorithms and Systems",
-	year	= "2015",
-	journal	= "Proceedings of the Very Large Datasets (VLDB)"
+	doi = {10.14778/2824032.2824132},
+	url = {https://doi.org/10.14778%2F2824032.2824132},
+	year = 2015,
+	month = {aug},
+	publisher = {{VLDB} Endowment},
+	volume = {8},
+	number = {12},
+	pages = {2040--2041},
+	author = {Arun Kejariwal and Sanjeev Kulkarni and Karthik Ramasamy},
+	title = {Real time analytics},
+	journal = {Proceedings of the {VLDB} {E}ndowment}
 }
-@article{ltc,
-	title={Lightweight temporal compression of microclimate datasets},
-	author={Schoellhammer, Tom and Greenstein, Ben and Osterweil, Eric and Wimbrow, Michael and Estrin, Deborah},
-	year={2004}
+@inproceedings{ltc,
+	doi = {10.1109/lcn.2004.72},
+	url = {https://doi.org/10.1109%2Flcn.2004.72},
+	publisher = {{IEEE} (Comput. Soc.)},
+	author = {T. Schoellhammer and B. Greenstein and E. Osterweil and M. Wimbrow and D. Estrin},
+	title = {Lightweight {T}emporal {C}ompression of {M}icroclimate {D}atasets [wireless sensor networks]},
+	booktitle = {29th Annual {IEEE} {I}nternational {C}onference on {L}ocal {C}omputer {N}etworks}
 }
-@article{ltcd,
-  title={A multi-dimensional extension of the Lightweight Temporal Compression method},
-  author={Li, Bo and Sarbishei, Omid and Nourani, Hosein and Glatard, Tristan},
-  journal={arXiv preprint arXiv:1811.09930},
-  year={2018}
+@inproceedings{ltcd,
+	doi = {10.1109/bigdata.2018.8621946},
+	url = {https://doi.org/10.1109%2Fbigdata.2018.8621946},
+	year = 2018,
+	month = {dec},
+	publisher = {{IEEE}},
+	author = {Bo Li and Omid Sarbishei and Hosein Nourani and Tristan Glatard},
+	title = {A multi-dimensional extension of the {L}ightweight {T}emporal {C}ompression method},
+	booktitle = {2018 {IEEE} {I}nternational {C}onference on {B}ig {D}ata (Big Data)}
 }
 @inproceedings{chained_reservoir_sampling,
   title={Sampling from a moving window over streaming data},
   author={Babcock, Brian and Datar, Mayur and Motwani, Rajeev},
-  booktitle={Proceedings of the thirteenth annual Association for Computing Machinery-SIAM symposium on Discrete algorithms},
+  booktitle={Proceedings of the thirteenth annual {A}ssociation for {C}omputing {M}achinery-{SIAM} {S}ymposium on {D}iscrete algorithms},
   pages={633--634},
   year={2002},
-  organization={Society for Industrial and Applied Mathematics}
+  organization={{S}ociety for {I}ndustrial and {A}pplied {M}athematics}
 }
 @inproceedings{cuckoo_filter,
-  title={Cuckoo filter: Practically better than bloom},
-  author={Fan, Bin and Andersen, Dave G and Kaminsky, Michael and Mitzenmacher, Michael D},
-  booktitle={Proceedings of the 10th Association for Computing Machinery International on Conference on emerging Networking Experiments and Technologies},
-  pages={75--88},
-  year={2014},
[…]
