### 2019-06-27 #1485 @gkthiruvathukal
@sepandhaghighi Many thanks for your excellent and thorough review efforts!

@azazel7  I have gone through the checklist, and I'm ready to accept.
Can you please archive the accepted version of the package and provide an associated DOI (e.g., via Zenodo)?
