### 2019-07-02 #1485 @azazel7
@gkthiruvathukal If I am not mistaken, the DOI is 10.5281/zenodo.3266109. The last release (1.0.0.1) is just for the paper modifications. The corresponding webpage is [there](https://zenodo.org/record/3266109).
