### 2019-07-22 #1485 @sepandhaghighi
@gkthiruvathukal
Hi
Just a friendly reminder.
I think you forgot this paper, its ready to be accepted by you.

Best Regards
Sepand Haghighi
