### 2019-07-23 #1485 @gkthiruvathukal
@sepandhaghighi Thanks for the reminder.

So here is the final checklist for @azazel7 to fill in. Can you please do the following? (This is something I am doing for all submissions now, so you can just check the ones you already did above.)

Once these are done, I will do a final proofread and signal the JOSS EICs to process acceptance.
