### 2019-07-25 #1485 @azazel7
@gkthiruvathukal

- [x] The tag release is [here](https://github.com/big-data-lab-team/OrpailleCC/releases).
- [x] The Zenodo archive is available [there](https://zenodo.org/record/3266109).
- [x] The metadata has been updated (author list with ORCID, the title of the release).
- [x] There is only one Zenodo DOI and it is 10.5281/zenodo.3266109.
