### 2019-07-25 #1485 @gkthiruvathukal
@openjournals/joss-editors  I recommend that we proceed with accepting this submission.
