### 2019-07-25 a2cdf6d
Update paper.bib
Fixes to a few references
