+++ paper/paper.bib
@@ -2,7 +2,7 @@
 	doi = {10.14778/2824032.2824132},
 	url = {https://doi.org/10.14778%2F2824032.2824132},
 	year = 2015,
-	month = {aug},
+	month = aug,
 	publisher = {{VLDB} Endowment},
 	volume = {8},
 	number = {12},
@@ -13,7 +13,7 @@
 }
 @inproceedings{ltc,
 	doi = {10.1109/lcn.2004.72},
-	url = {https://doi.org/10.1109%2Flcn.2004.72},
+	year = 2004,
 	publisher = {{IEEE} (Comput. Soc.)},
 	author = {T. Schoellhammer and B. Greenstein and E. Osterweil and M. Wimbrow and D. Estrin},
 	title = {Lightweight {T}emporal {C}ompression of {M}icroclimate {D}atasets [wireless sensor networks]},
@@ -150,12 +150,13 @@
 	journal = {{IEEE} {C}ommunications {M}agazine}
 }
 @misc{freertos,
-  title = {FreeRTOS},
+  title = {{FreeRTOS}},
+  author = {{Amazon Web Services}},
   howpublished = {\url{https://www.freertos.org/}},
   note = {Accessed: 2019-01-11}
 }
 @article{growing-iot,
-  title={IoT platforms: enabling the Internet of Things},
+  title={{IoT} platforms: enabling the {Internet of Things}},
   author={Lucero, Sam and others},
   journal={White paper},
   year={2016}
