### 2019-07-25 244a0c6
Update paper.md
Minor edit to the language around one reference
