### 2019-07-25 #1485 @kyleniemeyer
Hi @azazel7, could you please merge the two small PRs I just submitted, which make a few edits to your paper? https://github.com/big-data-lab-team/OrpailleCC/pull/15 and https://github.com/big-data-lab-team/OrpailleCC/pull/16
