### 2019-07-25 #1485 @kyleniemeyer
Congrats @azazel7 on your paper's publication in JOSS! Thanks to @sepandhaghighi for reviewing and @gkthiruvathukal for editing!
