### 2019-07-25 #1485 @whedon
:tada::tada::tada: Congratulations on your paper acceptance! :tada::tada::tada:
